modified modify-for-aws to first check if oski user exists

+++ post-install/modify-for-aws.sh
@@ -7,9 +7,11 @@
 # it will take a minute or two to complete
 
 # setup oski user with admin privileges and for passwordless SSH
-adduser oski sudo
-cp -pr /home/ubuntu/.ssh /home/oski
-chown -R oski:oski /home/oski/.ssh
+if [ ! -e /home/oski ]; then 
+    adduser oski sudo
+    cp -pr /home/ubuntu/.ssh /home/oski
+    chown -R oski:oski /home/oski/.ssh
+fi
 
 # add a few useful software packages
 apt-get update
